pyproject.toml
[project]
name = "openutm-verification"
version = "0.1.1"
description = "Verification tools for Flight Blender and OpenUTM products"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "httpx>=0.27.0",
    "arrow==1.3.0",
    "python-dotenv==1.0.1",
    "http-sfv==0.9.9",
    "cryptography==44.0.3",
    "jwt==1.3.1",
    "http-message-signatures==0.5.0",
    "redis==6.0.0",
    "walrus==0.9.4",
    "jwcrypto==1.5.6",
    "pyproj==3.7.1",
    "shapely==2.1.0",
    "implicitdict==3.0.0",
    "uas-standards==3.5.0",
    "geojson==3.2.0",
    "folium>=0.20.0",
    "faker==9.3.1",
    "pandas>=2.3.1",
    "pydeck>=0.9.1",
    "pythreejs>=2.4.2",
    "ipywidgets>=8.1.7",
    "numpy>=2.3.1",
    "pillow>=11.3.0",
    "earcut>=1.1.5",
    "pika>=1.3.2",
    "dacite>=1.9.2",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.6",
    "loguru>=0.7.2",
    "pydantic>=2.11.7",
]

[project.scripts]
openutm-verify = "openutm_verification.run_verification:main"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pylint-pytest>=1.1.7",
    "ruff>=0.12.10",
    "ty>=0.0.1a14",
    "pylint>=3.3.2",
    "pytest>=8.4.1",
    "mypy>=1.11.2",
    "types-pyyaml",
    "types-redis",
    "types-requests",
]

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/openutm_verification",
]

[tool.pytest.ini_options]
pythonpath = [
    ".",
    "src/openutm_verification",
]
testpaths = [
    "tests",
]

[tool.coverage.run]
source = [
    "src/openutm_verification",
]


[tool.pylint."messages control"]
disable = [
    "missing-module-docstring",
    "missing-function-docstring",
    "missing-class-docstring",
    "unused-import",
    "useless-import-alias",
    "invalid-name",
    "global-statement",
]

[tool.ruff]
line-length = 150
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
# select = ["E", "F", "W", "I", "N", "D", "B", "A", "C4", "S", "Q"]
# ignore = ["D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107", "S101"]
select = [
    "E",
    "F",
    "W",
    "I",
    # "UP",
]
ignore = [
    "F401",
]
fixable = [
    "I",
]
unfixable = [
]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.codespell]
ignore-words = ".codespellignore"
quiet-level = 2
skip = "uv.lock"

[tool.pylint]
load-plugins = "pylint_pytest"

[tool.pylint.format]
max-line-length = 150

[tool.mypy]
python_version = "3.12"
# warn_return_any = true
# warn_unused_configs = true
ignore_missing_imports = true
# disallow_untyped_defs = true
# check_untyped_defs = true

[tool.uv]
dev-dependencies = [
    "ruff",
    "pylint",
    "mypy",
    "types-pyyaml",
    "types-redis",
    "types-requests",
]
